Command for UserGrades & SelectSpeciality

+++ src/main/java/ua/training/admission/model/dao/GenericDao.java
@@ -9,7 +9,7 @@ import java.util.Optional;
  */
 public interface GenericDao<T> {
 	
-    Optional<T> findById(int id);
+    Optional<T> findById(Long id);
 
     List<T> findAll();
